Reponse reclamation

+++ main/database/migrations/2022_10_11_152746_create_reponse_reclamations_table.php
@@ -13,19 +13,20 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('reponse_reclamations', function (Blueprint $table) {
-            $table->id();
-            $table->text("reponse");
-            $table->dateTime("datereponse");
-            $table->foreignId('reclamation_id')
-                ->constrained()
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table->boolean('resolu')->nullable()->default(false);
-            $table->text('image')->nullable();
-        });
-    }
+        if(!Schema::hasTable('reponse_reclamations')) {
 
+            Schema::create('reponse_reclamations', function (Blueprint $table) {
+                $table->id();
+                $table->timestamps();
+                $table->text('reponse');
+                $table->foreignId('reclamations_id')
+                    ->constrained()
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table->boolean('resolu')->nullable()->default(false);
+                $table->text('image')->nullable();
+            });
+        }}
     /**
      * Reverse the migrations.
      *
